requirements.txt
fastapi
uvicorn
pydantic
scikit-image 
opencv-python 
numpy
pillow
